Refactor: isLoading, cartItems 조건분리 + Loader 컴포넌트 추가
